Move useful functions from the main to the package
